Fixed the boolean not.
